fix: rename delete user
